Add double click component functionality
When double clicking on a component in the FormDesigner sidebar, it will
automatically add that component to the bottom of the form. This
currently does not work for custom components and will show an error
notification saying that it is unimplemented for custom components. This
is because some custom components may require some data before the
component can even be added to the form. The normal method of doing this
is to drag a component into the form, a pop up modal will appear where
you can input that data, then you can click save to add that component
to the form.

Resolves: FORMS-4