Update columns and fix version number
The version number was incorrect because the versions we're not sorted
in descending order. The ManageVersions vue will retrieve the versions
from the API /forms/:id endpoint in descending order. The
formService.listVersions function does not sort it in descending order
by default.

The columns have been re-ordered and have a gutter separating the
actions and published version.