ExternalAPIs updated to composition API
ExternalAPIs is updated to the composition API. Form ID needed to be passed down instead of grabbing it from the form store. Likely issue is that one Vue component is resetting the form object because it's using the default value in ExternalAPIs. It's that or it is loading ExternalAPIs before getting form data.