Updating Cypress
Cypress updated to version 14.2.1
Moved files from the tests/functional directory to tests/functional/cypress
Removed cypress-drag-drop imports as they appear to not be in use.
Moved date-fns to devDependencies
Updated the devcontainers to install dependencies when the container is built.

TODO: Fix the tests.. they don't seem to work on 14.2.1